pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "base64url>=1.0.0",
    "bs4>=0.0.2",
    "fastapi[standard]>=0.128.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis[aioredis]>=7.1.0",
    "setuptools>=80.9.0",
    "uvicorn[standard]>=0.40.0",
    "numpy>=1.26.0",
    "soundfile>=0.12.1",
    "vonage>=4.7.2",
    "websockets>=15.0.1",
    "deepgram-sdk",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ruff>=0.14.10",
]
